requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
decorator==4.1.2
Django==2.1.2
djangorestframework==3.11.0
h5py==2.10.0
html5lib==0.9999999
httpie==2.0.0
idna==2.9
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Jinja2==2.11.1
jupyter-client==5.1.0
jupyter-core==4.3.0
Keras==2.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Mako==1.0.6
Markdown==2.2.0
MarkupSafe==1.0
matplotlib==2.0.2
mysqlclient==1.4.6
nose==1.3.7
numpy==1.13.1
pandas==0.20.3
path.py==10.3.1
pickleshare==0.7.4
prettytable==0.7.2
prompt-toolkit==1.0.15
protobuf==3.11.3
pyecharts==1.6.0
Pygments==2.6.1
pygpu==0.6.9
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==5.3
pyzmq==16.0.2
recurrentshop==1.0.0
requests==2.23.0
scikit-learn==0.19.0
scipy==0.19.1
simplegeneric==0.8.1
simplejson==3.17.0
six==1.10.0
tensorflow==1.2.0
Theano==0.9.0
tornado==4.5.2
traitlets==4.3.2
urllib3==1.25.8
wcwidth==0.1.7
Werkzeug==1.0.0
wincertstore==0.2
